# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Analyst API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CryptoPrice {
  id: string;
  symbol: string;
  name: string;
  current_price: number;
  price_change_24h: number;
  price_change_percentage_24h: number;
  market_cap: number;
  market_cap_rank: number;
  total_volume: number;
  high_24h: number;
  low_24h: number;
  image: string;
  ath: number;
  ath_change_percentage: number;
  circulating_supply: number;
}

export interface MarketOverview {
  total_market_cap: number;
  total_volume_24h: number;
  btc_dominance: number;
  eth_dominance: number;
  market_cap_change_24h: number;
  active_cryptocurrencies: number;
  fear_greed_index: number;
  fear_greed_label: string;
}

export interface PricePoint {
  timestamp: string;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface TrendingCoin {
  id: string;
  symbol: string;
  name: string;
  market_cap_rank: number;
  price_btc: number;
  score: number;
  image: string;
  price_change_percentage_24h: number;
  current_price_usd: number;
}

export interface WatchlistItem {
  id: number;
  symbol: string;
  name: string;
  added_at: string;
}

export interface AddWatchlistItem {
  symbol: string;
  name: string;
}

export interface SuccessMessage {
  message: string;
}

export type GetCryptoHistoryParams = {
  /**
   * Number of days of history (default 7)
   */
  days?: number;
};