# lib/api-zod/src/generated/types/addWatchlistItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface AddWatchlistItem {
  symbol: string;
  name: string;
}