# lib/api-zod/src/generated/types/marketOverview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface MarketOverview {
  total_market_cap: number;
  total_volume_24h: number;
  btc_dominance: number;
  eth_dominance: number;
  market_cap_change_24h: number;
  active_cryptocurrencies: number;
  fear_greed_index: number;
  fear_greed_label: string;
}