# lib/api-zod/src/generated/types/trendingCoin.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface TrendingCoin {
  id: string;
  symbol: string;
  name: string;
  market_cap_rank: number;
  price_btc: number;
  score: number;
  image: string;
  price_change_percentage_24h: number;
  current_price_usd: number;
}