# lib/api-zod/src/generated/types/watchlistItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface WatchlistItem {
  id: number;
  symbol: string;
  name: string;
  added_at: string;
}